Updated stat generation

+++ src/Airslip.Analytics.Core/Models/TimelyMetric.cs
@@ -2,6 +2,6 @@ using System.Collections.Generic;
 
 namespace Airslip.Analytics.Core.Models;
 
-public record TimelyMetric(int Period, string Description, long Balance, PeriodType PeriodType);
+public record TimelyMetric(int Period, string Description, double Balance, PeriodType PeriodType);
 
-public record Series(string Name, IEnumerable<TimelyMetric> Metrics, IEnumerable<long> Data);
+public record Series(string Name, IEnumerable<TimelyMetric> Metrics, IEnumerable<double> Data);
